xdp-tools-arago: Add version v1.5.4
This is a sideport and update from a recipe in meta-dpdk [1].  There are
no other version of this recipe in any of the standard locations
(oe-core, meta-oe), so we will carry this here for the time being.

There is a big "hack" fix with this recipe to get around an oe-core
multilib issue.  On arm and aarch64 some of the /usr/include/asm headers
are renamed and wrapped in the multilib system that uses the wordsize
for the platform (based on defines in the compiler), it chooses the
renamed file.  The issue is that the original file is renamed to either
-32.h or -64.h based on the target architecture, but only one of them is
created.  Usually this is ok, but for this package it presents a
problem.

Some of the files are built using clang with the -target bpf option to
build BPF programs.  In this case all of the defines are changed from
arm/aarch64 to just bpf, and this short circuits the multilib setup as
bpf is not checked in the wordsize.h that is installed by the multilib
flow.  The fallback wordsize is 32 bits in this case.  Since aarch64 is
64 bit, but the bpf is looking for the 32 bit headers we get compile
errors.

The hack solution is to copy the needed files from the sysroot into the
source for the program that the needed xx-32.h files are found.  There
is likely a better way to handle this in the long run, but for now this
allows us to build the package.

[1] https://git.yoctoproject.org/meta-dpdk/tree/dynamic-layers/clang-layer/recipes-support/xdp-tools?id=3493058e7dcf00c78517ec5efb6457ebb1e85089

Signed-off-by: Ryan Eatmon <reatmon@ti.com>